Updated server.js and code.js
Server:
Added functionality for favorites. Specifically, I added the routes "users/favorites/:username" and "users/add/favorites". The former is a GET request to retrieve the user's list of favorites, the latter is a POST request to add a recipe to their list of favorites.

Client:
Added getStrictMatchRecipes(), getRelaxedMatchedRecipes(), getAllRecipes(), searchByIngredient(), and searchRecipes() back to the client side code, as they seemed to have been deleted at some point. I also added two new functions getFavorites(), which makes a GET request to the server to get the user's favorites. The second is addFavorites(), which makes a POST request to the server to add a recipe to the user's favorites.